+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 ### Features
 
+- [#7685](https://github.com/blockscout/blockscout/pull/7685) - Add yul filter and "language" field for smart contracts
 - [#7653](https://github.com/blockscout/blockscout/pull/7653) - Add support for DEPOSIT and WITHDRAW token transfer event in older contracts
 - [#7628](https://github.com/blockscout/blockscout/pull/7628) - Support partially verified property from verifier MS; Add property to track contracts automatically verified via eth-bytecode-db
 - [#7603](https://github.com/blockscout/blockscout/pull/7603) - Add Polygon Edge and optimism genesis files support
@@ -14,6 +15,8 @@
 
 ### Fixes
 
+- [#7712](https://github.com/blockscout/blockscout/pull/7712) - Transaction actions import fix
+- [#7709](https://github.com/blockscout/blockscout/pull/7709) - Contract args displaying bug
 - [#7654](https://github.com/blockscout/blockscout/pull/7654) - Optimize exchange rates requests rate
 - [#7636](https://github.com/blockscout/blockscout/pull/7636) - Remove receive from read methods
 - [#7635](https://github.com/blockscout/blockscout/pull/7635) - Fix single 1155 transfer displaying
@@ -32,6 +35,7 @@
 
 ### Chore
 
+- [#7699](https://github.com/blockscout/blockscout/pull/7699) - Add block_number index for address_coin_balances table
 - [#7666](https://github.com/blockscout/blockscout/pull/7666) - Search label query
 - [#7644](https://github.com/blockscout/blockscout/pull/7644) - Publish docker images CI for prod/staging branches
 - [#7594](https://github.com/blockscout/blockscout/pull/7594) - Stats service support in docker-compose config with new frontend